Fix final fp map ack
Adds a new FPSendState enum value FINAL_ACK, which is used for finishing
the flightplan creation process between the sdk and autopilot.
Previously this step was handled in a "default" switch statement,
which caused an issue where the last step of the FP MAP process would
indefinitely send the final ack back and forth.