Split ac log on any packet w/ earlier sys_time
The SYS_INIT packet doesn't include sys_time, so it has to have that
field populated by the most recent packet that does contain sys_time.
This leads to a problem with splitting logs on SYS_INIT packets, since
they can contain at least 1 packet that has an earlier sys_time than the
rest of the log.

This updates the parser to split as soon as any packet is encountered
with an earlier sys_time value.